LayoutTests/imported/w3c/web-platform-tests/navigation-api/ordering-and-transition/currententrychange-before-popstate-intercept-expected.txt


FAIL currententrychange fires before popstate for navigation.back() and navigation.forward() assert_equals: expected 1 but got 0
